refactor: move all files outside of src directory inside custom snippets
