+++ src/extensionsIntegrated/CustomSnippets/UIHelper.js
