+++ src/extensionsIntegrated/CustomSnippets/global.js
