+++ src/extensionsIntegrated/CustomSnippets/helper.js
