+++ src/extensionsIntegrated/CustomSnippets/snippetCodeHints.js
