+++ src/extensionsIntegrated/CustomSnippets/snippetCursorManager.js
