+++ src/extensionsIntegrated/CustomSnippets/snippetsState.js
